etl/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
pymongo==4.6.1
neo4j==5.17.0
pandas==2.2.0
numpy==1.26.4
requests==2.31.0
